requirements.txt
gradio
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi 
uvicorn